requirements.txt
confluent-kafka
google-cloud-bigquery
requests
avro-python3